Fix: Implemented changes
