+++ app/mocks.ts
@@ -73,3 +73,166 @@ export const colors = [
   "bg-cyan-600",
   "bg-sky-600",
 ];
+
+
+export const transactions = [
+  {
+    id: "1",
+    date: "Today",
+    items: [
+      {
+          id: "t1",
+          type: "Swapped",
+          amount: "28.10",
+          currency: "USDC",
+          nativeValue: "NGN 369,723.88",
+          time: "4:14 AM",
+          status: "Completed",
+          fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+      },
+      {
+        id: "t2",
+        type: "Swapped",
+        amount: "15.50",
+        currency: "USDT",
+        nativeValue: "NGN 5,672.15",
+        time: "8:45 PM",
+        status: "Completed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+      },
+      {
+        id: "t3",
+        type: "Swapped",
+        amount: "100.00",
+        currency: "USDT",
+        nativeValue: "KES 45,678.90",
+        time: "2:20 PM",
+        status: "Failed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+      },
+    ],
+  },
+  {
+    id: "2",
+    date: "Yesterday",
+    items: [
+      {
+        id: "t4",
+        type: "Swapped",
+        amount: "28.10",
+        currency: "USDC",
+        nativeValue: "NGN 369,723.88",
+        time: "4:14 AM",
+        status: "Completed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+
+      },
+      {
+        id: "t5",
+        type: "Swapped",
+        amount: "15.50",
+        currency: "USDT",
+        nativeValue: "ARS 82,506.45",
+        time: "6:30 PM",
+        status: "Completed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+
+      },
+      {
+        id: "t6",
+        type: "Swapped",
+        amount: "200.00",
+        currency: "USDT",
+        nativeValue: "KES 7,890.65",
+        time: "8:00 AM",
+        status: "Completed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+
+      },
+      {
+        id: "t7",
+        type: "Swapped",
+        amount: "75.30",
+        currency: "USDC",
+        nativeValue: "NGN 53,674.20",
+        time: "5:55 PM",
+        status: "Failed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+
+      },
+    ],
+  },
+  {
+    id: "3",
+    date: "2 days ago",
+    items: [
+      {
+        id: "t8",
+        type: "Swapped",
+        amount: "28.10",
+        currency: "USDC",
+        nativeValue: "NGN 369,723.88",
+        time: "4:14 AM",
+        status: "Completed",
+        fees: "NGN 2,392",
+          recipient: "Francesca Tobiloba",
+          bank: "First Bank of Nigeria",
+          account: "0267856481",
+          memo: "From me, Donda North",
+          fundStatus: "Deposited",
+          timeSpent: "12 seconds",
+          day: "13 May 2024"
+
+      },
+    ],
+  },
+];
